refactor(ml): split model registry into a per-model package
The single MODELS tuple in model_registry.py would lengthen with every new
classifier. Split the module into a model_registry/ package: the
ModelDefinition dataclass and lifecycle status constants move to an internal
_model_definition.py, each built-in model's definition value moves to its own
definitions/<id>.py, and the package __init__ assembles the ordered MODELS
roster, hosts the lookup helpers, and runs the import-time invariant check.
Adding a classifier is now a new definition file plus a one-line roster entry.

Behavior-preserving: every existing import path, the
patch.object(model_registry, "MODELS", ...) retire seam, the import-time
validation, and the observable card order and default are unchanged, so the
full suite passes with no read-site edits.
