+++ app/astrodash/infrastructure/ml/model_registry/definitions/__init__.py
@@ -1,0 +1,7 @@
+"""Per-model definition values.
+
+Each module here defines exactly one built-in model as a
+:class:`~astrodash.infrastructure.ml.model_registry._model_definition.ModelDefinition`.
+Adding a built-in classifier is a new file here plus a one-line registration in
+the package root's ``MODELS`` roster.
+"""
